fix Prediction migration points column type and casts PHPDoc

+++ database/migrations/2026_06_10_190931_create_predictions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('fixture_id')->constrained()->cascadeOnDelete();
             $table->unsignedTinyInteger('home_score');
             $table->unsignedTinyInteger('away_score');
-            $table->unsignedTinyInteger('points')->nullable();
+            $table->unsignedSmallInteger('points')->nullable();
             $table->timestamps();
 
             $table->unique(['user_id', 'fixture_id']);
